@@ -1,5 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
+/// <reference types="react" />
+
 /**
  * 该文件由 Tailchat 自动生成
  * 用于插件的类型声明
@@ -66,6 +68,8 @@ declare module '@capital/common' {
 
   export const appendUrlSearch: any;
 
+  export const getServiceWorkerRegistration: any;
+
   export const useGroupIdContext: any;
 
   export const getServiceUrl: () => string;
@@ -84,7 +88,14 @@ declare module '@capital/common' {
 
   export const getCachedConverseInfo: any;
 
-  export const localTrans: any;
+  /**
+   * 本地翻译
+   * @example
+   * localTrans({'zh-CN': '你好', 'en-US': 'Hello'});
+   *
+   * @param trans 翻译对象
+   */
+  export const localTrans: (trans: Record<'zh-CN' | 'en-US', string>) => string;
 
   export const getLanguage: any;
 
@@ -114,7 +125,7 @@ declare module '@capital/common' {
 
   export const useLocation: any;
 
-  export const useHistory: any;
+  export const useNavigate: any;
 
   export const createFastFormSchema: any;
 
@@ -242,9 +253,14 @@ declare module '@capital/component' {
 
   export const Image: any;
 
-  export const Icon: any;
+  export const Icon: React.FC<{ icon: string } & React.SVGProps<SVGSVGElement>>;
 
-  export const IconBtn: any;
+  export const IconBtn: React.FC<{
+    icon: string;
+    iconClassName?: string;
+    shape?: 'circle' | 'square';
+    title?: string;
+  }>;
 
   export const PillTabs: any;
 
@@ -296,4 +312,6 @@ declare module '@capital/component' {
     userId: string;
     className?: string;
   }>;
+
+  export const Markdown: any;
 }
